Break helper into its own file
